Restructured archiving core into submodule
