+++ yark/archiver/reporter.py
@@ -3,7 +3,7 @@
 from colorama import Fore, Style
 import datetime
 from .video import Video, Element
-from .utils import _truncate_text
+from ..utils import _truncate_text
 from typing import TYPE_CHECKING, Optional
 
 if TYPE_CHECKING:
